+++ src/definition/ublox_qzss_svid_prn_map.h
@@ -5,9 +5,21 @@
 // Entries       : 4
 // Strategy      : switch
 
+#if defined(__AVR__)
+#include "../internal/avr_std/cstdint"
+#else
 #include <cstdint>
+#endif
+#if defined(__AVR__)
+#include "../internal/avr_std/optional"
+#else
 #include <optional>
+#endif
+#if defined(__AVR__)
+#include "../internal/avr_std/string_view"
+#else
 #include <string_view>
+#endif
 #include "../azaraC_config.h"
 #include "../internal/FlashString.h"