+++ src/framer/Frame.h
@@ -1,7 +1,11 @@
 #pragma once
 // azaraC - src/framer/Frame.h
 
+#if defined(__AVR__)
+#include "../internal/avr_std/cstdint"
+#else
 #include <cstdint>
+#endif
 
 namespace azaraC {
 namespace internal {