+++ src/framer/IFramer.h
@@ -1,7 +1,11 @@
 #pragma once
 // azaraC - src/framer/IFramer.h
 
+#if defined(__AVR__)
+#include "../internal/avr_std/cstdint"
+#else
 #include <cstdint>
+#endif
 #include "Frame.h"
 
 namespace azaraC {