+++ src/json/JsonWriter.h
@@ -3,8 +3,16 @@
 // Common JSON writer helpers for serializers
 
 #include "../Message.h"
+#if defined(__AVR__)
+#include "../internal/avr_std/optional"
+#else
 #include <optional>
+#endif
+#if defined(__AVR__)
+#include "../internal/avr_std/string_view"
+#else
 #include <string_view>
+#endif
 
 #if defined(ARDUINO) && ARDUINO >= 1
 #  include <Print.h>